Card Displayed Without Shrinking
I was having problems earlier where the card would shrink to fit it's content perfectly, but it's now solved due to setting the flex-grow to 1, so that it takes up as much vertical space as the height of the card not the height of the content. Next, I will need to have the cards display on queue when the buttons are clicked.