Fixed Overload Bug
There was a bug where after having randomly generated every card, the web page would crash. It turns out this was because I was forgetting to clear the array of stored indices that represented the cards that had already been chosen. This is now fixed, and the random generation works for as long as needed now.